Move common components to features
