+++ src/pages/_app.tsx
@@ -15,7 +15,7 @@ import {
 import { far } from "@fortawesome/free-regular-svg-icons";
 import { AppProps } from "next/dist/next-server/lib/router/router";
 import { initFirebase } from "src/utils/firbase";
-import { GlobalStyle } from "src/components/global-style";
+import { GlobalStyle } from "src/features/global-style";
 
 library.add(fab, fas, far, faHome, faPlus, faSignInAlt, faTimes, faStar, faCog);
 initFirebase();
